+++ .gitignore
@@ -1,5 +1,6 @@
-testing.ipynb
+sample.ipynb
 __pycache__
 /fetch_data/options/__pycache__
 sample.py
-/temp
+/temporary
+/dev_utils
