+++ core/data/nse/nse_index_symbols.json
