+++ core/data/nse/nse_stock_symbols.json
