+++ core/data/smart_api/bse_symbols.json
