+++ core/data/smart_api/nse_symbols.json
