+++ core/routers/smart_api/utils/constants.py
