+++ core/schemas/stock_model.py
